requirements.txt
dbt-core
dbt-postgres
requests
python-dotenv
